New terminal is ready
